mushbounds_clamped_size: add necessary casts

+++ bounds.all.c
@@ -11,10 +11,12 @@ size_t mushbounds_size(const mushbounds* bounds) {
    return sz;
 }
 size_t mushbounds_clamped_size(const mushbounds* bounds) {
-   size_t sz = mush_size_t_add_clamped(bounds->end.v[0] - bounds->beg.v[0], 1);
+   size_t sz = mush_size_t_add_clamped(
+      (size_t)bounds->end.v[0] - (size_t)bounds->beg.v[0], 1);
    for (mushdim i = 1; i < MUSHSPACE_DIM; ++i)
       sz = mush_size_t_mul_clamped(
-         sz, mush_size_t_add_clamped(bounds->end.v[i] - bounds->beg.v[i], 1));
+         sz, mush_size_t_add_clamped(
+            (size_t)bounds->end.v[i] - (size_t)bounds->beg.v[i], 1));
    return sz;
 }
